Create main layout
The main layout has two different versions for mobile devices and
desktop computers.

The mobile version has a main navigation in two rows, sidebar
navigation that is hidden by default, and the rest of the screen is
taken by the main content. There is a sandwich button to display the
idebar navigation on demand.

The desktop version is divided into three sections. A main navigation
ontop, a sidebar navigation on the left of the screen and a main
content section taking the rest of the space. The sandwich button is
hidden, because the sidebar navigation is shown by default and is not
collapsable.

When you resize the screen the layout switches from the mobile version
to the desktop and vice versa.