Create Sanbox page and refactor styles
The sanbox page is made of two rows with different cards.
It has two layouts. One for mobile devices and one for desktop
computers.

Large style refactoring.
The style.css file was getting bigger and bigger, and it was getting
hard to navigate in it. I decided to split it into "components". There
are 5 components: cards, main content layout, main navigation, page
layout and sidebar navigation.